Added AJAX refreshing on the MyAlerts lsiting. It's not an automatic refresh in that it requires the clocking of a link. I will be adding an auto refresh soon though.

+++ inc/plugins/MyAlerts/Alerts.class.php
@@ -134,7 +134,7 @@ class Alerts
 			$alerts  = "'".my_strtolower(implode("','", $alerts))."'";
 		}
 
-		return $this->db->update_query('alerts', array('unread' => '0'), 'id IN('.$alerts.')');
+		return $this->db->update_query('alerts', array('unread' => '0'), 'id IN('.$alerts.') AND uid = '.$this->mybb->user['uid']);
 	}
 
 	/**
@@ -150,7 +150,7 @@ class Alerts
 			$alerts  ="'".my_strtolower(implode("','", $alerts))."'";
 		}
 
-		return $this->db->delete_query('alerts', 'id IN('.$alerts.')');
+		return $this->db->delete_query('alerts', 'id IN('.$alerts.') AND uid = '.$this->mybb->user['uid']);
 	}
 
 	/**
